Fix firehose issues (#44)
These are similar fixes to what has been done in the streams plugin.

Please see the commit d0db8d4 as it looks like the same problem from #41 but I think this is a slightly neater way of handling it.

We've been running this branch for 3 weeks now and it's gone from crashing very often to not crashing at all.

Closes #39

Changelog:

* Properly format json for firehose

This was already done for streams in a086eea4

* Fix panic on Firehose ack/retry

Not entirely sure what the problem is but we've seen panics from multiple places in the code. Mostly copying changes that were made to the streams client in ce91e040fbf1c2ee9385465665e7e1fe8a8fa6d9 and hoping it helps.

* Fix nil dereference in Firehose failed responses

The test fails with the old function and passes with the new one. I haven't seen the actual responses from the API but I suspect that when some records failed and others passed, only the ones that failed have an ErrorCode. So it needs to check if `r.ErrorCode != nil` before checking the value. It seems the `aws.StringValue` helper function does that and also removes the need for the other nil check.